requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
pymongo==3.11.1
pyrebase4==4.4.1
Flask-JWT==0.3.2
firebase_admin==4.4.0
google-auth==1.23.0
requests==2.25.0
dnspython==1.16.0